+++ tests/test_likelihoods_lba.py
@@ -144,24 +144,3 @@ def test_lba3_basic():
 
 
 test_lba3_basic()
-
-
-def test_lba_sampling():
-    """Test if sampling works for available lba models."""
-    lba2_data_out = hssm.simulate_data(
-        model="lba2", theta=dict(A=0.2, b=0.5, v0=1.0, v1=1.0), size=500
-    )
-
-    lba3_data_out = hssm.simulate_data(
-        model="lba3", theta=dict(A=0.2, b=0.5, v0=1.0, v1=1.0, v2=1.0), size=500
-    )
-
-    lba2_model = hssm.HSSM(model="lba2", data=lba2_data_out)
-
-    lba3_model = hssm.HSSM(model="lba3", data=lba3_data_out)
-
-    traces_2 = lba2_model.sample(sampler="nuts_numpyro", draws=100, tune=200, chains=1)
-    traces_3 = lba3_model.sample(sampler="nuts_numpyro", draws=100, tune=200, chains=1)
-
-    assert isinstance(traces_2, az.InferenceData)
-    assert isinstance(traces_3, az.InferenceData)
